move test_provider_configuration
